backend-master/src/requirements_shared.txt
Flask==1.0.2
requests==2.21.0
flask_cors==3.0.2
flask-restplus==0.12.1
SQLAlchemy==1.2.15
flask_sqlalchemy==2.3.2
Flask-Migrate==2.3.1
PyYaml==3.13
pytest==4.0.2
pytest-cov==2.6.0
pylint==2.3.1
requests-mock==1.5.2
mysqlclient==1.3.14
Flask-Caching==1.4.0
redis-py-cluster==1.3.6
Flask-Cache-Redis-Cluster==0.0.5
deepdiff==3.3.0
pydatadeck
elastic-apm[flask]==3.0.1